# packages/api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Diff
# during dev, should point to your local machine
host: localhost:8080
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /authenticate:
    # binds a127 app logic to a route
    x-swagger-router-controller: routes
    post:
      description: Submits a comment
      # used as the method name of the controller
      operationId: authenticate
      consumes:
        - "application/octet-stream"
        - "text/plain"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UserToken"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
...
definitions:
  UserToken:
    required:
      - access_token      
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string  